# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/anime": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Anime"
                ],
                "summary": "Call All Data Anime",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Anime"
                ],
                "summary": "Add New Anime",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Anime"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/anime/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Anime"
                ],
                "summary": "Call Single Data Anime",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search anime by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Anime"
                ],
                "summary": "Update Single Data Anime",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update anime by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.AnimeRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Anime"
                ],
                "summary": "Delete Anime",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete anime",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/family": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Family"
                ],
                "summary": "Call All Data Family",
                "responses": {}
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Family"
                ],
                "summary": "Add New Family",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Family"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/family/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Family"
                ],
                "summary": "Call Single Data Family",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search family by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Family"
                ],
                "summary": "Update Single Data Family",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update family by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.FamilyRequest"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Family"
                ],
                "summary": "Delete Family",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete family",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Anime": {
            "type": "object",
            "properties": {
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "episode": {
                    "type": "integer"
                },
                "family_id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                },
                "season": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.AnimeRequest": {
            "type": "object",
            "properties": {
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "episode": {
                    "type": "integer"
                },
                "family_id": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer"
                },
                "season": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.Family": {
            "type": "object",
            "properties": {
                "alt_name": {
                    "type": "string"
                },
                "avg_rate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_duration": {
                    "type": "integer"
                }
            }
        },
        "models.FamilyRequest": {
            "type": "object",
            "properties": {
                "alt_name": {
                    "type": "string"
                },
                "avg_rate": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total_duration": {
                    "type": "integer"
                }
            }
        }
    }
}